Split travel schema & document into own file
